requirements.txt
# You need these requirements in order to contribute to this reusable app.
django==1.4.1
django-libs==0.6
south==0.7.6
fabric==1.4.3
factory_boy==1.2.0
django-nose==1.1
coverage==3.5.2
django-coverage==1.2.2
mock==1.0b1
selenium==2.25.0
django-jasmine==0.3.2
